requirements.txt
netCDF4 ~= 1.6.3
numpy ~= 1.24.2
python-cmr ~= 0.9.0
